+++ src/test/java/org/alfasoftware/soapstone/external/TestExternalCreation.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.alfasoftware.soapstone.integration;
+package org.alfasoftware.soapstone.external;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.alfasoftware.soapstone.SoapstoneServiceBuilder;
